[TASK] controller structure refactoring
+ hashtag handling in frontend redone
